[DD-953] refactored PropertyGraphGDMReader
