+++ src/main/java/org/dswarm/graph/gdm/read/GDMResourceReader.java
@@ -19,9 +19,7 @@ package org.dswarm.graph.gdm.read;
 import org.dswarm.graph.DMPGraphException;
 import org.dswarm.graph.json.Resource;
 
-public interface GDMResourceReader {
+public interface GDMResourceReader extends GDMReader {
 
 	public Resource read() throws DMPGraphException;
-
-	public long countStatements();
 }
