+++ src/main/java/org/dswarm/graph/gdm/read/PropertyGraphGDMResourceByIDReader.java
@@ -19,6 +19,8 @@ package org.dswarm.graph.gdm.read;
 import org.dswarm.common.model.AttributePath;
 import org.dswarm.graph.DMPGraphException;
 import org.dswarm.graph.delta.util.GraphDBUtil;
+
+import com.google.common.base.Optional;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Node;
 import org.slf4j.Logger;
@@ -31,14 +33,16 @@ public class PropertyGraphGDMResourceByIDReader extends PropertyGraphGDMResource
 
 	private static final Logger	LOG	= LoggerFactory.getLogger(PropertyGraphGDMResourceByIDReader.class);
 
+	private static final String type = "GDM record by ID";
+
 	private final String		recordId;
 	private String recordURI;
 	private final AttributePath	recordIdentifierAP;
 
-	public PropertyGraphGDMResourceByIDReader(final String recordIdArg, final AttributePath recordIdentifierAPArg, final String dataModelUri,
+	public PropertyGraphGDMResourceByIDReader(final String recordIdArg, final AttributePath recordIdentifierAPArg, final String dataModelUri, final Optional<Integer> optionalVersionArg,
 			final GraphDatabaseService database) throws DMPGraphException {
 
-		super(dataModelUri, database);
+		super(dataModelUri, optionalVersionArg, database, type);
 
 		recordId = recordIdArg;
 		recordIdentifierAP = recordIdentifierAPArg;
@@ -46,7 +50,7 @@ public class PropertyGraphGDMResourceByIDReader extends PropertyGraphGDMResource
 	}
 
 	@Override
-	protected Node getResourceNode() {
+	protected Node getResourceNode() throws DMPGraphException {
 
 		if (recordURI == null) {
 
@@ -55,7 +59,7 @@ public class PropertyGraphGDMResourceByIDReader extends PropertyGraphGDMResource
 			return null;
 		}
 
-		final PropertyGraphGDMResourceByURIReader uriReader = new PropertyGraphGDMResourceByURIReader(recordURI, dataModelUri, database);
+		final PropertyGraphGDMResourceByURIReader uriReader = new PropertyGraphGDMResourceByURIReader(recordURI, dataModelUri, Optional.of(version), database);
 
 		return uriReader.getResourceNode();
 	}
