+++ src/main/java/org/dswarm/graph/gdm/read/PropertyGraphGDMResourceByURIReader.java
@@ -16,8 +16,11 @@
  */
 package org.dswarm.graph.gdm.read;
 
+import org.dswarm.graph.DMPGraphException;
 import org.dswarm.graph.GraphIndexStatics;
 import org.dswarm.graph.model.GraphStatics;
+
+import com.google.common.base.Optional;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.index.Index;
@@ -28,17 +31,20 @@ import org.neo4j.graphdb.index.IndexHits;
  */
 public class PropertyGraphGDMResourceByURIReader extends PropertyGraphGDMResourceReader {
 
+	private static final String type = "GDM record by URI";
+
 	private final String	recordUri;
 
-	public PropertyGraphGDMResourceByURIReader(final String recordUriArg, final String dataModelUri, final GraphDatabaseService database) {
+	public PropertyGraphGDMResourceByURIReader(final String recordUriArg, final String dataModelUri, final Optional<Integer> optionalVersionArg, final GraphDatabaseService database)
+			throws DMPGraphException {
 
-		super(dataModelUri, database);
+		super(dataModelUri, optionalVersionArg, database, type);
 
 		recordUri = recordUriArg;
 	}
 
 	@Override
-	protected Node getResourceNode() {
+	protected Node getResourceNode()  throws DMPGraphException {
 
 		final Index<Node> resourcesWDataModel = database.index().forNodes(GraphIndexStatics.RESOURCES_W_DATA_MODEL_INDEX_NAME);
 		final IndexHits<Node> hits = resourcesWDataModel.get(GraphStatics.URI_W_DATA_MODEL, recordUri + dataModelUri);
